uvcvideo: Remo OBSBOT quirk fix for incorrect relative min pan/tilt/zoom
The OBSBOT series of cameras misreports the minimum relative pan_speed, tilt_speed and zoom_continuous v4l2 controls resulting in the inability to control the camera with finesse with an analog stick or other programmatic methods. This patch applies to all Remo (OBSBOT) vendor cameras with the vendor ID of 0x3564 all of which exhibit this behavior. If the vendor fixes the firmware this behavior should still remain valid but the kernel won't need to be modified for new models when they are released.

Signed-off-by: John Bauer <johnebgood@securitylive.com>